requirements.txt
banal==1.0.1
normality==2.1.1
pantomime==0.4.1
followthemoney==2.1.9
followthemoney-store[postgresql]==2.2.3
servicelayer[google,amazon]==1.15.1
languagecodes==1.0.7
countrytagger==0.1.2
pyicu==2.5
google-cloud-vision==2.0.0
tesserocr==2.5.1
spacy==2.3.2
fingerprints==1.0.1
fasttext==0.9.2

# Development
nose==1.3.7
coverage
click==7.1.2

# File format support
dbf==0.99.0
pdflib==0.3.0
pdfminer.six
pymediainfo==4.2.1
python-magic==0.4.18
pypdf2==1.26.0
rarfile==4.0
xlrd==1.2.0
openpyxl==3.0.5
odfpy==1.4.1
cchardet==2.1.6
lxml==4.6.1
olefile==0.46
pillow==8.0.0
vobject==0.9.6.1
msglite==0.29.0

cryptography==3.1.1
requests[security]==2.24.0
